[voicecall] Refactored telepathy plugin to split channel type handling.
